+++ src/test/java/com/github/alexisjehan/mavencheck/core/component/session/MavenSessionTest.java
@@ -50,7 +50,7 @@ final class MavenSessionTest {
 
 	@Test
 	void testConstructorInvalid() {
-		try (final var mockedMavenUtils = Mockito.mockStatic(MavenUtils.class)) {
+		try (var mockedMavenUtils = Mockito.mockStatic(MavenUtils.class)) {
 			mockedMavenUtils.when(MavenUtils::makeSettings)
 					.thenThrow(SettingsBuildingException.class);
 			assertThatExceptionOfType(MavenSessionException.class)
@@ -64,8 +64,10 @@ final class MavenSessionTest {
 		final var remoteRepositories = List.of(MavenUtils.createRemoteRepository("foo-id", "https://foo-host"));
 		Mockito.when(mockedRepositorySystem.newResolutionRepositories(Mockito.any(), Mockito.any()))
 				.thenReturn(remoteRepositories);
-		try (final var mockedMavenUtils = Mockito.mockStatic(MavenUtils.class)) {
-			mockedMavenUtils.when(() -> MavenUtils.makeRepositorySystem(Mockito.any()))
+		try (var mockedMavenUtils = Mockito.mockStatic(MavenUtils.class)) {
+			mockedMavenUtils.when(MavenUtils::makeServiceLocator)
+					.thenCallRealMethod();
+			mockedMavenUtils.when(() -> MavenUtils.makeRepositorySystem(Mockito.notNull()))
 					.thenReturn(mockedRepositorySystem);
 			final var mavenSession = new MavenSession();
 			assertThat(mavenSession.resolve(List.of())).isSameAs(remoteRepositories);
@@ -84,8 +86,10 @@ final class MavenSessionTest {
 		final var versionRangeResult = new VersionRangeResult(new VersionRangeRequest());
 		Mockito.when(mockedRepositorySystem.resolveVersionRange(Mockito.any(), Mockito.any()))
 				.thenReturn(versionRangeResult);
-		try (final var mockedMavenUtils = Mockito.mockStatic(MavenUtils.class)) {
-			mockedMavenUtils.when(() -> MavenUtils.makeRepositorySystem(Mockito.any()))
+		try (var mockedMavenUtils = Mockito.mockStatic(MavenUtils.class)) {
+			mockedMavenUtils.when(MavenUtils::makeServiceLocator)
+					.thenCallRealMethod();
+			mockedMavenUtils.when(() -> MavenUtils.makeRepositorySystem(Mockito.notNull()))
 					.thenReturn(mockedRepositorySystem);
 			final var mavenSession = new MavenSession();
 			assertThat(mavenSession.request(new VersionRangeRequest())).isSameAs(versionRangeResult);